CODE: activity ratio can be a float

+++ examples/activeBD.py
@@ -146,5 +146,5 @@ if __name__ == "__main__":
             "This script takes in 2 arguments: [gpuidi (int)], [activity_ratio (float)]"
         )
     gpuid = int(sys.argv[1])
-    activity_ratio = int(sys.argv[2])
+    activity_ratio = float(sys.argv[2])
     run_monomer_diffusion(gpuid, N, ids, activity_ratio)
